Moved ConsoleLog, BinderResolver to Flame.Front
In addition, dsc handles unknown extensions more gracefully.
