+++ dsc/Projects/ProjectHandlers.cs
@@ -1,4 +1,7 @@
-﻿using Flame.Front.Projects;
+﻿using Flame.Compiler;
+using Flame.Compiler.Projects;
+using Flame.Front;
+using Flame.Front.Projects;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -34,6 +37,29 @@ namespace dsc.Projects
                 }
             }
             return null;
+        }
+
+        public static IProjectHandler GetProjectHandler(ProjectPath Path)
+        {
+            var handler = ProjectHandlers.GetHandler(Path);
+            if (handler == null)
+            {
+                if (string.IsNullOrWhiteSpace(Path.Extension))
+                {
+                    ConsoleLog.Instance.LogError(new LogEntry("Invalid project extension", "'" + Path.Path.Path + "' does not have an extension."));
+                }
+                else
+                {
+                    ConsoleLog.Instance.LogError(new LogEntry("Invalid project extension", "Extension '" + Path.Extension + "' in '" + Path.Path.Path + "' was not recognized as a known project extension."));
+                }
+                ConsoleLog.Instance.WriteLine("Supported extensions:");
+                foreach (var item in handlers.SelectMany(item => item.Extensions))
+                {
+                    ConsoleLog.Instance.WriteLine(" *." + item, ConsoleColor.Yellow);
+                }
+                throw new NotSupportedException();
+            }
+            return handler;
         }
     }
 }
